requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
asgiref==3.8.1
atlastk==0.13.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.2
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
priority==2.0.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.2
Quart==0.19.6
tortoise==0.1.1
tortoise-orm==0.20.1
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.2.1
yarl==1.16.0
Flask-Caching==2.3.0
cachelib==0.9.0
